844. Backspace String Compare
